fix: Pin outline window to the buffer
This resolves the situtaion which allows external plugin or for the user
to open the file buffer within the outline window. E.g. here are a few
scenarios I've run into:

https://github.com/user-attachments/assets/1c0fb255-bb56-49cd-9299-76b56be1b6ec

With the new options applied it is not working exactly as I want to (I'd
like buffer to be openning in the oil.nvim place) but it at least
doesn't enter the broken state and doesn't create the empty OUTLINE_1
file

https://github.com/user-attachments/assets/dfa0335a-9972-40ac-afca-f8f16f867818